requirements.txt
# TeachCE CRM Integration Dependencies

# Environment and configuration
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# HubSpot Official SDK
hubspot-api-client>=8.0.0

# Excel handling
openpyxl>=3.1.2

# Data manipulation
pandas>=2.0.0

# Logging
python-json-logger>=2.0.0
